.BUGFIX unreachable health check for ALB target group
Since the health check of Fabio is used to check the target group
health and hence give the ALB the possibility to forward traffic to
the instances -> port 9998 has to be open for ingress.

Refactored the naming of the security group as well. Since it reflects
more then just the docker port security group configuration.