.NEW grafana preconfiguration - prometheus datasource, dashboard
NOTE: The automatic loading of the dashboard is not working!
I tried many this to find the issue. The steps described in the manual
does not seem to work. There might be a bug in Grafana itself. The
current state is documented in the README.md.
Manual import and export of dashboards is nevertheless working fine
and the currently used approach. An example is provided: dasboard_cos.json.