.NEW basic Elasticsearch and Kibana setup
Very basic setup of Elasticsearch and Kibana 6.2.3 as separate
container but in the same nomad job group. This is not ideal but
allows to automatically configure the Elasticsearch URL for
Kibana.

NOTE: Using Fabio to access the services does not work because of the
prefix path! Because of this problem ALB access and using the internal
load balancer is not working. The services have to be accessed
directly using the client ip and the dynamically assigned port.