.NEW Activate privileged mode for nomad client nodes - DANGEROUS!
Some services might need to run Docker container in privileged
mode. For example the buidsystem Concourse. To facilitate the handling
for this demo setup one AMI with activated privileged mode is
generate.
Nevertheless activating the privileged mode in general for all client
nodes is dangerous. For the future the terraform module has to be
updated to use different AMIs for different datacenters,
e.g. activating privileged mode support for backoffice and deactivate
it for public-services.