.NEW first steps to separate logging service setup
If all parts of the logging are handled in one group they will be
placed on one instance. Furthermore it is not possible to scale the
container number separately. Hence the components have to be placed in
separate job description, but then the interconnection of the
components becames more hard - currently unsolved.