.UPDATE Test Consul Example WORKING + Validity checks for cluster setup
The consul cluster is deployed and it is check if there is a leader elected and
if there are active members in the cluster.
This should be sufficient to verify a working cluster setup.

Furthermore a helper script to export the generated private SSH key was added to
facilitate debugging of the tests.