.REFACTOR moved examples into separate files
The main test file examples_test.go became quite big.
To improve readability and separation of concerns each test of an example
moved into a separate file.
Furthermore the helper functionality was migrated into a separate file as
well. This file now contains the main control parameters for the tests.